cleaning up structure and vestigial junk
