+++ man/fparse.Rd
@@ -57,9 +57,9 @@ fload(
   }
 }}
 
-\item{query}{If not \code{NULL}, a string used as a JSON Pointer to identify a
-specific element within \code{json}.
-\code{character(1L)}, default: \code{NULL}}
+\item{query}{If not \code{NULL}, JSON Pointer(s) used to identify and extract
+specific elements within \code{json}. See Details and Examples.
+\code{NULL}, \code{character()}, or \code{list()} of \code{character()}. default: \code{NULL}}
 
 \item{empty_array}{Any R object to return for empty JSON arrays.
 default: \code{NULL}}
@@ -106,7 +106,7 @@ default: \code{NULL}.}
 \itemize{
   \item \code{"double"} or \code{0L}: big integers become \code{double}s
   \item \code{"string"} or \code{1L}: big integers become \code{character}s
-  \item \code{"integer64"} or \code{2L}: big integers \code{bit64::integer64}s
+  \item \code{"integer64"} or \code{2L}: big integers become \code{bit64::integer64}s
 }}
 
 \item{verbose}{Whether to display status messages.
@@ -141,6 +141,25 @@ Parse JSON strings and files to R objects.
           returned object will have the same names.
     \item If \code{json} contains multiple values and is unnamed, \code{fload()}
           names each returned element using the file's \code{basename()}.
+   }
+
+   \item \code{query}'s goal is to minimize te amount of data that must be
+   materialized as R objects (the main performance bottleneck) as well as
+   facilitate any post-parse processing.
+   \itemize{
+     \item To maximize flexibility, there are two approaches to consider when designing \code{query} arguments.
+     \itemize{
+       \item \code{character} vectors are interpreted as containing queries that
+       meant to be applied to all elements of \code{json=}.
+       \itemize{
+         \item If \code{json=} contains 3 strings and \code{query=} contains
+         3 strings, the returned object will be a list of 3 elements (1 for each element
+         of \code{json=}), which themselves each contain 3 lists (1 for each element
+         of \code{query=}).
+       }
+       \item \code{list}s of \code{character} vectors are interpreted as containing
+       queries meant to be applied  to \code{json} in a zip-like fashion.
+     }
    }
 
 }
